+++ src/common/webviews/views/aurafx.ts
@@ -1,14 +1,14 @@
 import { getGenericConfig } from '@common/utils/configutils';
 
-import { externalToolOpened as openExternalTool, externalToolWarning } from '../common';
+import { externalToolWarning, openToolInternally } from '../common';
 
-const aurafxLink = 'https://aurafx.vercel.app/';
+const aurafxLink = 'https://aurafx.vercel.app?source=MythicScribe';
 const name = 'AuraFX by Sleepsweety';
 
 export function openAuraFXWebview(): void {
     if (!!getGenericConfig('allowExternalTools') === false) {
         return externalToolWarning(aurafxLink);
     }
-    openExternalTool(name, aurafxLink);
+    openToolInternally(name, aurafxLink);
     return;
 }